check if dataverse to dataverse links already exist #7259
