+++ src/main/java/propertyFiles/Bundle.properties
@@ -769,6 +769,7 @@ dataverse.saved.search.failure=The saved search was not able to be linked.
 dataverse.linked.success= {0} has been successfully linked to {1}.
 dataverse.linked.success.wait= {0} has been successfully linked to {1}. Please wait for its contents to appear.
 dataverse.linked.internalerror={0} has been successfully linked to {1} but contents will not appear until an internal error has been fixed.
+dataverse.linked.error.alreadyLinked={0} has already been linked to {1}.
 dataverse.page.pre=Previous
 dataverse.page.next=Next
 dataverse.byCategory=Dataverses by Category
