+++ src/test/java/edu/harvard/iq/dataverse/api/DataversesIT.java
@@ -383,40 +383,9 @@ public class DataversesIT {
         }
 
     }
-    
-    @Test
-    public void testCreateDeleteDataverseLink() {
-        Response createUser = UtilIT.createRandomUser();
-        
-        createUser.prettyPrint();
-        String username = UtilIT.getUsernameFromResponse(createUser);
-        String apiToken = UtilIT.getApiTokenFromResponse(createUser);
-        
-        Response superuserResponse = UtilIT.makeSuperUser(username);
-        
-        Response createDataverseResponse = UtilIT.createRandomDataverse(apiToken);
-        createDataverseResponse.prettyPrint();
-        String dataverseAlias = UtilIT.getAliasFromResponse(createDataverseResponse);
-        Integer dataverseId = UtilIT.getDataverseIdFromResponse(createDataverseResponse);
-        
-        Response createDataverseResponse2 = UtilIT.createRandomDataverse(apiToken);
-        createDataverseResponse2.prettyPrint();
-        String dataverseAlias2 = UtilIT.getAliasFromResponse(createDataverseResponse2);
-        
-        Response createLinkingDataverseResponse = UtilIT.createDataverseLink(dataverseAlias, dataverseAlias2, apiToken);
-        createLinkingDataverseResponse.prettyPrint();
-        
-        createLinkingDataverseResponse.then().assertThat()
-                .body("data.message", equalTo("Dataverse " + dataverseAlias + " linked successfully to " + dataverseAlias2))
-                .statusCode(200);
-        
-        Response deleteLinkingDataverseResponse = UtilIT.deleteDataverseLink(dataverseAlias, dataverseAlias2, apiToken);
-        deleteLinkingDataverseResponse.prettyPrint();
-        deleteLinkingDataverseResponse.then().assertThat()
-                .body("data.message", equalTo("Link from Dataverse " + dataverseAlias + " to linked Dataverse " + dataverseAlias2 + " deleted"))
-                .statusCode(200);
-    }
-    
+
+    // testCreateDeleteDataverseLink was here but is now in LinkIT
+
     @Test
     public void testUpdateDefaultContributorRole() {
         Response createUser = UtilIT.createRandomUser();
