+++ src/test/java/edu/harvard/iq/dataverse/api/LinkIT.java
@@ -101,4 +101,42 @@ public class LinkIT {
                 .body("message", equalTo("Can't link a dataset that has already been linked to this dataverse"));
     }
 
+    @Test
+    public void testCreateDeleteDataverseLink() {
+        Response createUser = UtilIT.createRandomUser();
+
+        createUser.prettyPrint();
+        String username = UtilIT.getUsernameFromResponse(createUser);
+        String apiToken = UtilIT.getApiTokenFromResponse(createUser);
+
+        Response superuserResponse = UtilIT.makeSuperUser(username);
+
+        Response createDataverseResponse = UtilIT.createRandomDataverse(apiToken);
+        createDataverseResponse.prettyPrint();
+        String dataverseAlias = UtilIT.getAliasFromResponse(createDataverseResponse);
+        Integer dataverseId = UtilIT.getDataverseIdFromResponse(createDataverseResponse);
+
+        Response createDataverseResponse2 = UtilIT.createRandomDataverse(apiToken);
+        createDataverseResponse2.prettyPrint();
+        String dataverseAlias2 = UtilIT.getAliasFromResponse(createDataverseResponse2);
+
+        Response createLinkingDataverseResponse = UtilIT.createDataverseLink(dataverseAlias, dataverseAlias2, apiToken);
+        createLinkingDataverseResponse.prettyPrint();
+        createLinkingDataverseResponse.then().assertThat()
+                .statusCode(OK.getStatusCode())
+                .body("data.message", equalTo("Dataverse " + dataverseAlias + " linked successfully to " + dataverseAlias2));
+
+        Response tryLinkingAgain = UtilIT.createDataverseLink(dataverseAlias, dataverseAlias2, apiToken);
+        tryLinkingAgain.prettyPrint();
+        tryLinkingAgain.then().assertThat()
+                .statusCode(FORBIDDEN.getStatusCode())
+                .body("message", equalTo(dataverseAlias + " has already been linked to " + dataverseAlias2 + "."));
+
+        Response deleteLinkingDataverseResponse = UtilIT.deleteDataverseLink(dataverseAlias, dataverseAlias2, apiToken);
+        deleteLinkingDataverseResponse.prettyPrint();
+        deleteLinkingDataverseResponse.then().assertThat()
+                .statusCode(OK.getStatusCode())
+                .body("data.message", equalTo("Link from Dataverse " + dataverseAlias + " to linked Dataverse " + dataverseAlias2 + " deleted"));
+    }
+
 }
